Initial commit of MLLM evaluation scripts
All required libraries to run the models are provided in the 'requirements.txt' file.

We are initially releasing the inference scripts for the GPT-based models. Each script is configured with task-specific instructions, and the models are provided with in-context learning samples directly in the prompt.

Please ensure that the few-shot and test set data are properly separated into the directory structure required by the scripts.

API keys for this research were managed using a .env file with the python-dotenv library.

These scripts were used to conduct the experiments with GPT-4o and GPT-5.

Scripts for the open-source models and the full evaluation code will be released in a future update.